requirements.txt
python-dotenv
seleniumbase
requests